# Optimize `Array` `min`/`max` methods