Updated `Array::min` and `Array::max` to not copy Variants for comparison, and store index instead of copying each time new min/max is found.

Makes min/max around 30% faster for ints, and around 80% faster for strings, compared with gdscript below:

```gdscript
func _ready() -> void:
	test_min(range(100), 100000, "min int")
	test_min(range(100).map(str), 100000, "min str")
	test_max(range(100), 100000, "max int")
	test_max(range(100).map(str), 100000, "max str")

func test_min(p_array : Array, p_times : int, p_name : String):
	var start := Time.get_ticks_msec()
	for i in p_times:
		p_array.min()
	var end := Time.get_ticks_msec()
	print("%s: %dms" % [p_name, end - start])
	
func test_max(p_array : Array, p_times : int, p_name : String):
	var start := Time.get_ticks_msec()
	for i in p_times:
		p_array.max()
	var end := Time.get_ticks_msec()
	print("%s: %dms" % [p_name, end - start])
```

old
```
min int: 230ms
min str: 396ms
max int: 231ms
max str: 428ms
```

new
```
min int: 169ms
min str: 228ms
max int: 171ms
max str: 228ms
```

<!--
Please target the `master` branch in priority.

Relevant fixes are cherry-picked for stable branches as needed by maintainers.

To speed up the contribution process and avoid CI errors, please set up pre-commit hooks locally:
https://docs.godotengine.org/en/latest/contributing/development/code_style_guidelines.html
-->